requirements.txt
python-telegram-bot
flask
google-genai
python-dotenv
